refactor(sqlite): fix encapsulation in atomic/ module
- Remove drop_guard.rs: move Drop impl into session.rs (Rust convention)
- Make AtomicWriteSession fields private (remove pub(super))
- Add AtomicWriteSession::close_without_rename() (test-only) so writer.rs
  no longer accesses session fields directly (Tell, Don't Ask)
- Extract retry_policy.rs: RetryPolicy trait + implementations
- Extract constants.rs: SQLITE_SIDECAR_SUFFIXES
- Extract tests.rs: move tests out of mod.rs
- mod.rs reduced to declarations and re-exports only (~35 lines)
